List of languages
Languages are listed as they appear in the sources (WOLD and LDS). We have 3 Guaycuruan and  all Mataguayan languages. The list also includes 2 Enhlet-Enhlet lgs, as well as Guaraní and one Zamucoan language. There are some inconsistencies according to what we currently know in terms of  the name employed to refer to some languages, e.g., Sanapaná.